feat: 待办索引扫描与 todos 表扩展
**索引模块（todo/index.rs）**
- scan_vault：遍历 vault 内所有 md 文件，跳过 .noteidea/ 和 assets/
- 跳过代码块内的形似待办行（三反引号围栏）
- 单文件失败只跳过该文件并返回错误清单，不中断整体扫描
- list_tags：按使用频次降序返回全库已有标签
- 10 个单元测试覆盖：空 vault、单文件多待办、代码块屏蔽、
  文件读取失败、删库重扫一致性、标签频次排序

**数据库扩展（db.rs）**
- todos 表增加 time_expr/recurrence/intensity 三列存储提醒元数据
- 迁移逻辑：ALTER TABLE ADD COLUMN（SQLite 3.2.0+ 已支持）
- user_version 升至 2
- integrity_check 在升级后自动执行

**其他**
- config.rs：default_remind_time 配置项，缺失或非法时回落 09:00
- syntax.rs：serialize 与 marker_at 标记 #[allow(dead_code)]
- Cargo.toml：新增 walkdir 和 tempfile(dev) 依赖

对应 spec：todo/index 任务 8.1-8.4
测试：70 项全绿，clippy 干净

+++ src-tauri/src/lib.rs
@@ -106,7 +106,7 @@ fn parse_todo_line(text: String) -> Option<todo::syntax::TodoLine> {
 
 /// 查询 vault 内已有标签，按使用频次排序（供 # 弹层）
 #[tauri::command]
-fn list_tags(db: State<'_, db::Handle>) -> Result<Vec<(String, u32)>, String> {
+fn list_tags(db: State<'_, db::Handle>) -> Result<Vec<(String, usize)>, String> {
     todo::index::list_tags(&db)
 }
